Update kazoo.py
I need to change this later, the while loop is all wrong. Likely blow this up into three separate functions, one for starting the music during a caution, one for stopping the music when "one to green" is seen, and one to toggle between the two.

However, it at least does play music.