Init sample rate fix
Turns out the mixer would only quit once the flag showed one to green, which means songs with different sample rates would still play improperly if they were picked. For example, "foo.mp3" was selected, and the mixer was initialized at 44.1Khz. Once "foo.mp3" was finished playing, the script selected "bar.mp3", which had a sample rate of 48Khz, but because the mixer re-initialized, it plays at the wrong sample rate, making the audio sound weird.

Throwing in an "if False" check for "mixer.music.get_busy()" and unloading the mixer should fix this problem.