Fixed potential error after song finishes
Fixed a mistake at the end of the cautionMusic() function, that would try to quit the mixer after it had already been uninitialized. This change will only allow this additional quit to be called if the mixer is not busy, but is still initialized (only happens if we are still under caution).